separate a buy post list to an offer and a like
