+++ src/components/shop/PostList/OfferTabPostList/types.ts
@@ -1,0 +1,8 @@
+import type { OfferSummary, Review } from '@types'
+
+export type OfferPostListProps = {
+  className?: string
+  posts: OfferSummary[]
+  onClickWriteReview(offer: OfferSummary): void
+  onClickReadReview(review: Review): void
+}
