+++ src/components/shop/PostList/index.tsx
@@ -1,3 +1,4 @@
-export * from './BuyTabPostList'
+export * from './OfferTabPostList'
+export * from './LikeTabPostList'
 export * from './SaleTabPostList'
 export * from './ReviewTabPostList'
