+++ src/pages/shop/view/buy/index.tsx
@@ -1,26 +1,18 @@
 import { SelectBox } from '@offer-ui/react'
 import type { ReactElement } from 'react'
 import { useState } from 'react'
+import { LikePanelView } from './panel/like'
+import { OfferPanelView } from './panel/offer'
 import { Styled } from './styled'
-import { useModal } from '@hooks/useModal'
-import {
-  useGetMyOffersQuery,
-  useGetLikedPostsQuery,
-  useUpdateLikeStatusMutation,
-  useReviewsMutation
-} from '@apis'
-import type { ReviewState } from '@components'
-import { Tabs, BuyTabPostList, ReviewModal } from '@components'
+import { useGetMyOffersQuery, useGetLikedPostsQuery } from '@apis'
+import { Tabs } from '@components'
 import { TRADE_ACTIVITY_TYPES, SORT_OPTIONS } from '@constants'
 import type {
-  OfferSummary,
-  Review,
   SortOption,
   SortOptionCodes,
   TradeBuyActivityCodes,
   TradeBuyActivityNames
 } from '@types'
-import { isNumber } from '@utils'
 
 const tradeBuyActivityList = Object.entries<
   TradeBuyActivityCodes,
@@ -32,54 +24,12 @@ export const ShopPageBuyView = (): ReactElement => {
     useState<SortOptionCodes>('CREATED_DATE_DESC')
   const [activityType, setActivityType] =
     useState<TradeBuyActivityCodes>('like')
-  const [readReviewContent, setReadReviewContent] = useState<Review>(
-    {} as Review
-  )
-  const [writeReviewContent, setWriteReviewContent] = useState<OfferSummary>(
-    {} as OfferSummary
-  )
-
-  const readReviewModal = useModal()
-  const writeReviewModal = useModal()
 
   const offers = useGetMyOffersQuery({ sort: sortOptionCode })
   const likedPosts = useGetLikedPostsQuery({ sort: sortOptionCode })
-  const likeStatusMutation = useUpdateLikeStatusMutation()
-  const reviewsMutation = useReviewsMutation()
 
   const handleChangeSortOption = (newSortOption: SortOption) => {
     setSortOptionCode(newSortOption.code)
-  }
-  const handleChangeActivityType =
-    (newActivityType: TradeBuyActivityCodes) => (): void =>
-      setActivityType(newActivityType)
-
-  const handleChangeProductLikeStatus = async (postId: number) => {
-    await likeStatusMutation.mutateAsync({ postId })
-    likedPosts.refetch()
-  }
-
-  const handleOpenReadReview = (review: Review) => {
-    setReadReviewContent(review)
-    readReviewModal.openModal()
-  }
-  const handleOpenWriteReview = (offer: OfferSummary) => {
-    setWriteReviewContent(offer)
-    writeReviewModal.openModal()
-  }
-  const handleConfirmWriteReview = async (reviewState: ReviewState) => {
-    if (!isNumber(reviewState.reviewScore)) {
-      return
-    }
-
-    await reviewsMutation.mutateAsync({
-      targetMemberId: writeReviewContent.seller.id,
-      postId: writeReviewContent.postId,
-      score: reviewState.reviewScore,
-      content: reviewState.reviewText
-    })
-    await offers.refetch()
-    writeReviewModal.closeModal()
   }
 
   return (
@@ -95,7 +45,7 @@ export const ShopPageBuyView = (): ReactElement => {
                 return (
                   <Styled.Tab
                     key={`${code}-tab`}
-                    onClick={handleChangeActivityType(code)}>
+                    onClick={() => setActivityType(code)}>
                     <Styled.StatusButtonLabel>
                       <Styled.Circle isCurrent={isCurrent} />
                       <Styled.StatusButton isCurrent={isCurrent} type="button">
@@ -120,36 +70,14 @@ export const ShopPageBuyView = (): ReactElement => {
           </Styled.SearchOptionsWrapper>
           <Styled.ProductListPanels>
             <Tabs.Panel>
-              <BuyTabPostList
-                activityType="like"
-                posts={likedPosts.data?.posts || []}
-                onChangeProductLikeStatus={handleChangeProductLikeStatus}
-              />
+              <LikePanelView sortOptionCode={sortOptionCode} />
             </Tabs.Panel>
             <Tabs.Panel>
-              <BuyTabPostList
-                activityType="offer"
-                posts={offers.data?.offers || []}
-                onClickReadReview={handleOpenReadReview}
-                onClickWriteReview={handleOpenWriteReview}
-              />
+              <OfferPanelView sortOptionCode={sortOptionCode} />
             </Tabs.Panel>
           </Styled.ProductListPanels>
         </Tabs>
       </Styled.UserProductsWrapper>
-      <ReviewModal.Read
-        {...readReviewContent}
-        isOpen={readReviewModal.isOpen}
-        onClose={readReviewModal.closeModal}
-        onConfirm={readReviewModal.closeModal}
-      />
-      <ReviewModal.Write
-        isOpen={writeReviewModal.isOpen}
-        nickname={writeReviewContent.seller?.nickname}
-        productName={writeReviewContent.title}
-        onClose={writeReviewModal.closeModal}
-        onConfirm={handleConfirmWriteReview}
-      />
     </div>
   )
 }
